requirements.txt
# Install in Codespaces (postCreateCommand), local venv, or container — see CODESPACES.md
google-adk
fastapi
uvicorn[standard]
google-cloud-firestore
google-api-python-client
google-auth
google-auth-oauthlib
python-dotenv
